refactor: modified the components to jsx extensions
