+++ src/components/privateRoutes/PrivateRoute.jsx
@@ -1,9 +1,9 @@
-import { Route, Navigate } from "react-router-dom";
+import { Navigate, Outlet } from "react-router-dom";
 export const PrivateRoute = ({ auth, element, path }) => {
   return (
     <>
       {auth ? (
-        <Route path={path} element={element} />
+        <Outlet />
       ) : (
         <Navigate to="/login" state={{ from: path }} replace />
       )}
